+++ extensions/authChallenges.php
@@ -1,0 +1,7 @@
+<?php
+
+use Beebmx\KirbyCourier\CourierChallenge;
+
+return [
+    'courier' => CourierChallenge::class,
+];