+++ tests/Feature/ArquitectureTest.php
@@ -1,0 +1,5 @@
+<?php
+
+arch('globals')
+    ->expect(['dd', 'dump', 'ray'])
+    ->not->toBeUsed();